Adds arguments to the vagrant reload plugin
